git-svn-id: http://svn.terracotta.org/svn/ehcache@5933 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache-core-2.6.x/src/main/java/net/sf/ehcache/store/disk/DiskStorageFactory.java
@@ -124,17 +124,16 @@ public class DiskStorageFactory {
         this.diskStorePathManager = cache.getCacheManager().getDiskStorePathManager();
         this.file = diskStorePathManager.getFile(cache.getName(), ".data");
 
-        if (diskStorePathManager.isAutoCreated()) {
+        this.indexFile = diskStorePathManager.getFile(cache.getName(), ".index");
+        this.pinningEnabled = determineCachePinned(cache.getCacheConfiguration());
+        this.diskPersistent = cache.getCacheConfiguration().isDiskPersistent();
+
+        if (diskPersistent && diskStorePathManager.isAutoCreated()) {
             LOG.warn("Data in persistent disk stores is ignored for stores from automatically created directories.\n"
                     + "Remove diskPersistent or resolve the conflicting disk paths in cache configuration.\n" + "Deleting data file "
                     + file.getAbsolutePath());
             deleteFile(file);
-        }
-        this.indexFile = diskStorePathManager.getFile(cache.getName(), ".index");
-        this.pinningEnabled = determineCachePinned(cache.getCacheConfiguration());
-
-        diskPersistent = cache.getCacheConfiguration().isDiskPersistent();
-        if (!diskPersistent) {
+        } else if (!diskPersistent) {
             deleteFile(file);
             deleteFile(indexFile);
         }
